requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
sentry-sdk==0.16.5
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
